+++ packages/main/test/specs/Menu.spec.js
@@ -160,6 +160,20 @@ describe("Menu interaction", () => {
 			assert.strictEqual(await busyIndicator.getProperty("size"), "Medium", "Size attribute is properly set.");
 			assert.strictEqual(await busyIndicator.getProperty("delay"), 100, "Delay attribute is properly set.");
 		});
+
+		it("Prevent menu closing on item press", async () => {
+			await browser.url(`test/pages/Menu.html`);
+			const openButton = await browser.$("#btnOpen");
+			openButton.click();
+			await browser.pause(100);
+
+			const menuPopover = await browser.$("ui5-static-area-item:last-of-type").shadow$("ui5-responsive-popover");
+			const newFileItem = await menuPopover.$("ui5-li[accessible-name='New File']");
+			newFileItem.click();
+			await browser.pause(100);
+
+			assert.ok(await menuPopover.getProperty("open"), "Menu is still opened.");
+		});
 	});
 
 describe("Menu Accessibility", () => {