+++ packages/playground/_stories/fiori/NotificationListItem/NotificationListItem.stories.ts
@@ -13,8 +13,8 @@ const component = "ui5-li-notification";
 
 export default {
 	title: "Fiori/Notification List Item",
-	component,
-	subcomponents: { 'NotificationAction': 'ui5-notification-action' },
+	component: "NotificationListItem",
+	subcomponents: { 'NotificationAction': 'NotificationAction' },
 	parameters: {
 		docs: {
 			page: DocsPage({ ...componentInfo, component }),