+++ packages/playground/_stories/fiori/ProductSwitch/ProductSwitch.stories.ts
@@ -15,8 +15,8 @@ const component = "ui5-product-switch";
 
 export default {
     title: "Fiori/ProductSwitch",
-    component,
-    subcomponents: { ProductSwitchItem: "ui5-product-switch-item" },
+    component: "ProductSwitch",
+    subcomponents: { ProductSwitchItem: "ProductSwitchItem" },
     parameters: {
         docs: {
             page: DocsPage({ ...componentInfo, component }),