+++ packages/playground/_stories/fiori/ShellBar/ShellBar.stories.ts
@@ -15,8 +15,8 @@ const component = "ui5-shellbar";
 
 export default {
     title: "Fiori/ShellBar",
-    component,
-    subcomponents: { ShellBarItem: "ui5-shellbar-item" },
+    component: "ShellBar",
+    subcomponents: { ShellBarItem: "ShellbarItem" },
     parameters: {
         docs: {
             page: DocsPage({ ...componentInfo, component }),