+++ packages/playground/_stories/main/TabContainer/TabContainer.stories.ts
@@ -13,8 +13,8 @@ const component = "ui5-tabcontainer";
 
 export default {
 	title: "Main/Tab Container",
-	component,
-	subcomponents: { "Tab": "ui5-tab", "TabSeparator": "ui5-tab-separator" },
+	component: "TabContainer",
+	subcomponents: { "Tab": "Tab", "TabSeparator": "TabSeparator" },
 	parameters: {
 		docs: {
 			page: DocsPage({ ...componentInfo, component })